AWS mocks

+++ manager/src/main/scala/io/hydrosphere/serving/manager/service/modelsource/S3ModelSource.scala
@@ -4,11 +4,9 @@ import java.io.{File, FileNotFoundException}
 import java.net.URI
 import java.nio.file.{Files, Path, Paths, StandardCopyOption}
 
-import awscala.s3.S3
 import io.hydrosphere.serving.manager.{LocalModelSourceConfiguration, S3ModelSourceConfiguration}
 import org.apache.logging.log4j.scala.Logging
-
-import scala.collection.concurrent.TrieMap
+import scala.collection.JavaConversions._
 
 /**
   * Created by Bulat on 31.05.2017.
@@ -18,27 +16,36 @@ class S3ModelSource(val configuration: S3ModelSourceConfiguration) extends Model
   private[this] val fileCache = new LocalModelSource(
     LocalModelSourceConfiguration(s"proxy-${configuration.name}", localFolder)
   )
-  private[this] implicit val s3 = S3.at(configuration.region)
-  private[this] val bucketObj = s3.bucket(configuration.bucket).get
 
-  private def downloadObject(objectPath: String): File = {
+  val client = configuration.s3Client
+
+  val proxyFolder: Path = Paths.get(localFolder)
+
+  def deleteProxyObject(objectPath: String): Boolean = {
+    Files.deleteIfExists(Paths.get(localFolder, objectPath))
+  }
+
+  def downloadObject(objectPath: String): File = {
     logger.debug(s"downloadObject: $objectPath")
-    val fileStream = bucketObj.getObject(objectPath).getOrElse(throw new FileNotFoundException(objectPath)).content
+    if (!client.doesObjectExist(configuration.bucket, objectPath))
+      throw new FileNotFoundException(objectPath)
+    val fileStream = client.getObject(configuration.bucket, objectPath).getObjectContent
 
     val folderStructure = Paths.get(localFolder, objectPath.split("/").dropRight(1).mkString("/"))
     Files.createDirectories(folderStructure)
 
     val file = Files.createFile(Paths.get(localFolder, objectPath))
     Files.copy(fileStream, file, StandardCopyOption.REPLACE_EXISTING)
-    val f = file.toFile
-    println(f)
-    f
+    file.toFile
   }
 
   override def getSubDirs(path: String): List[String] = {
     logger.debug(s"getSubDirs: $path")
     val pUri = Paths.get(path).toUri
-    bucketObj.keys(path)
+    client
+      .listObjects(configuration.bucket)
+      .getObjectSummaries
+      .map(_.getKey)
       .map(Paths.get(_).toUri)
       .map(k => pUri.relativize(k).toString.split("/").head)
       .distinct.toList
@@ -46,19 +53,33 @@ class S3ModelSource(val configuration: S3ModelSourceConfiguration) extends Model
 
   override def getSubDirs: List[String] = {
     logger.debug(s"getSubDirs")
-    bucketObj.keys().map(_.split("/").head).distinct.toList
+    val r = client
+      .listObjects(configuration.bucket)
+      .getObjectSummaries
+      .map(_.getKey)
+      .map(_.split("/").head)
+      .distinct
+      .toList
+    println(r)
+    r
   }
 
   override def getAllFiles(folder: String): List[String] = {
     logger.debug(s"getAllFiles: $folder")
     val pUri = URI.create(s"$folder")
-    val modelKeys = bucketObj.keys(pUri.toString).toList
+    val modelKeys = client
+      .listObjects(configuration.bucket, pUri.toString)
+      .getObjectSummaries
+      .map(_.getKey)
+    println(modelKeys)
     modelKeys
-      .map(k => URI.create(k))
-      .map(k => pUri.relativize(k))
-      .map(k => getReadableFile(k.toString))
+      .map(URI.create)
+      .map(_.toString)
+      .map(getReadableFile)
 
-    fileCache.getAllFiles(folder)
+    val r = fileCache.getAllFiles(folder)
+    println(r)
+    r
   }
 
   override def getReadableFile(path: String): File = {
